Added a functionality to the add button, so the Notes are stored in the database.

See merge request cse1105/2024-2025/teams/csep-team-35!26

+++ commons/src/test/java/commons/ServerTest.java
@@ -21,8 +21,8 @@ public class ServerTest {
     @Test
     public void getterAndSetterTest() {
         server1.setServerId(3);
-        int expected = 3;
-        int actual = server1.getServerId();
+        long expected = 3;
+        long actual = server1.getServerId();
         assertEquals(expected, actual);
     }
